+++ src/main/kotlin/racing/InputView.kt
@@ -1,0 +1,9 @@
+package racing
+
+interface InputView {
+    fun askCarCount()
+
+    fun askTryCount()
+
+    fun inputCount(count: String?): Int
+}
